Fix dest path for copying music
I had previously only tested `-copy=NONE`. My earlier fix to the return statements here was mistaken - it needs to return the path to the song file, and destinationPath was only the directory.

Tested manuall with each of the four copy options. NONE still works with the source file paths. The others all provide relative paths to the music, rooted at the provided `-output` path.